+++ src/components/QuestionCard/QuestionCard.tsx
@@ -158,6 +158,7 @@ const QuestionCard = ({
                     onShare={onShare}
                     onBookmark={onBookmark}
                     marked={marked}
+                    isAnswered={showAnswer}
                 />
 
                 <div className="p-4 sm:p-6">